Motor holder fastening
Added correct length from center of motor holder
to edge of cube. Added holes in correct place. As
well

The thickness of the cube will probably have to be adjusted.

I do not have the energy to find out how to flip them in
the assembly. The "cube" is pointing the wrong way.